third_party/WebKit/LayoutTests/fast/mediacapturefromelement/CanvasCaptureMediaStream-capture-out-of-DOM-element.html

<!DOCTYPE html>
<script src =../../resources/testharness.js></script>
<script src =../../resources/testharnessreport.js></script>
<script>

async_test(t => {
  var canvas = document.createElement('canvas');

  var recorder = new MediaRecorder(canvas.captureStream());
  recorder.ondataavailable = function() {
    t.step_func_done(function() {
      assert_true(event.data.size > 0, 'Recorded data size should be > 0');
    })();
    recorder.stop();
  }
  recorder.start(0);

  var ctx = canvas.getContext('2d');
  ctx.fillStyle = 'green';
  ctx.fillRect(0, 0, canvas.width, canvas.height);
}, "Verify that drawing to a 2D canvas that is not attached to the DOM dispatches a frame to an attached MediaRecorder." );

async_test(t => {
  var canvas = document.createElement('canvas');

  var recorder = new MediaRecorder(canvas.captureStream());
  recorder.ondataavailable = function() {
    t.step_func_done(function() {
      assert_true(event.data.size > 0, 'Recorded data size should be > 0');
    })();
    recorder.stop();
  }
  recorder.start(0);

  var gl = canvas.getContext('webgl');
  gl.clearColor(0, 1, 0, 1);
  gl.clear(gl.COLOR_BUFFER_BIT);
}, "Verify that drawing to a webgl canvas that is not attached to the DOM dispatches a frame to an attached MediaRecorder." );
</script>
